third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-palette-values-invalid-expected.txt

This is a testharness.js-based test.
FAIL CSS Fonts Module Level 4: parsing @font-palette-values assert_equals: expected 19 but got 0
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 1 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 2 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 3 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 4 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 5 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 6 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 7 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 8 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 9 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 10 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 11 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 12 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 13 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 14 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 15 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 16 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 17 Cannot read properties of undefined (reading 'cssText')
PASS CSS Fonts Module Level 4: parsing @font-palette-values 18
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 19 Cannot read properties of undefined (reading 'cssText')
FAIL CSS Fonts Module Level 4: parsing @font-palette-values 20 Cannot read properties of undefined (reading 'cssText')
Harness: the test ran to completion.
